% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.array.mlx}
\alias{as.array.mlx}
\title{Convert MLX array to R array}
\usage{
\method{as.array}{mlx}(x, ...)
}
\arguments{
\item{x}{An mlx array.}

\item{...}{Additional arguments (ignored)}
}
\value{
A numeric array.
}
\description{
Convert MLX array to R array
}
\examples{
x <- mlx_matrix(1:8, 2, 4)
as.array(x)
}
