% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.logical.mlx}
\alias{as.logical.mlx}
\title{Convert MLX array to logical vector}
\usage{
\method{as.logical}{mlx}(x, ...)
}
\arguments{
\item{x}{An mlx array.}

\item{...}{Additional arguments passed to \code{\link[base:vector]{base::as.vector()}}.}
}
\value{
A logical vector.
}
\description{
Convert MLX array to logical vector
}
\examples{
x <- as_mlx(c(1, 0, 2))
as.logical(x)
}
