% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{cbind.mlx}
\alias{cbind.mlx}
\title{Column-bind MLX tensors}
\usage{
\method{cbind}{mlx}(..., deparse.level = 1)
}
\arguments{
\item{...}{Objects to bind. MLX tensors are kept in MLX; other inputs are
coerced via \code{as_mlx()}.}
}
\value{
An \code{mlx} tensor stacked along the second axis.
}
\description{
Column-bind MLX tensors
}
