% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{colMeans.mlx}
\alias{colMeans.mlx}
\title{Column means of MLX matrix}
\usage{
\method{colMeans}{mlx}(x, na.rm = FALSE, dims = 1, ...)
}
\arguments{
\item{x}{An \code{mlx} matrix}

\item{na.rm}{Ignored (for compatibility)}

\item{dims}{Which dimensions are regarded as columns (default: 1)}

\item{...}{Additional arguments (ignored)}
}
\value{
An \code{mlx} vector with column means
}
\description{
Column means of MLX matrix
}
