% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.R
\name{compare_runs}
\alias{compare_runs}
\title{Compare OFV and parameters of selected runs}
\usage{
compare_runs(runnums, path = ".", ifOFV = TRUE, ifParam = TRUE)
}
\arguments{
\item{runnums}{\verb{<chr>} Partial model file names of the model files}

\item{path}{\verb{<chr>} Path to model directory for \link{get_mlx}}

\item{ifOFV}{\verb{<lgl>} \code{TRUE} to return objective function values}

\item{ifParam}{\verb{<lgl>} \code{TRUE} to return parameter values}
}
\value{
A list containing the OFV and parameter values of all models
}
\description{
Compare OFV and parameters of selected runs
}
\examples{
\dontrun{
# Model files are matched by regular expression 
compare_runs("/r01|/r02") # model names include "/r01" and "/r02"
compare_runs("^r01|^r02") # model names starting with "r01" and "r02"
}
}
