% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{conv_params}
\alias{conv_params}
\title{Common Convolution Parameters}
\arguments{
\item{input}{Input mlx array. Shape depends on dimensionality (see individual functions).}

\item{weight}{Weight array. Shape depends on dimensionality (see individual functions).}

\item{stride}{Stride of the convolution. Can be a scalar or vector (length depends
on dimensionality). Default: 1 for 1D, c(1,1) for 2D, c(1,1,1) for 3D.}

\item{padding}{Amount of zero padding. Can be a scalar or vector (length depends
on dimensionality). Default: 0 for 1D, c(0,0) for 2D, c(0,0,0) for 3D.}

\item{dilation}{Spacing between kernel elements. Can be a scalar or vector (length
depends on dimensionality). Default: 1 for 1D, c(1,1) for 2D, c(1,1,1) for 3D.}

\item{groups}{Number of blocked connections from input to output channels. Default: 1.}
}
\description{
Common Convolution Parameters
}
\keyword{internal}
