% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{crossprod.mlx}
\alias{crossprod.mlx}
\title{Cross product}
\usage{
\method{crossprod}{mlx}(x, y = NULL, ...)
}
\arguments{
\item{x}{An mlx matrix (2-dimensional array).}

\item{y}{An mlx matrix (default: NULL, uses x)}

\item{...}{Additional arguments forwarded to the corresponding base R implementation for signature compatibility.}
}
\value{
\code{t(x) \%*\% y} as an mlx object.
}
\description{
Cross product
}
\examples{
x <- mlx_matrix(1:6, 2, 3)
crossprod(x)
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.matmul}{mlx.core.matmul}
}
