% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{diag}
\alias{diag}
\title{Diagonal matrix extraction and construction}
\usage{
diag(x = 1, nrow, ncol, names = TRUE)
}
\arguments{
\item{x}{An object.}

\item{nrow, ncol}{Optional dimensions for matrix construction.}

\item{names}{Logical indicating whether to use names.}
}
\description{
Generic function for extracting/constructing diagonal matrices.
}
