% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim<-.mlx}
\alias{dim<-.mlx}
\title{Set dimensions of MLX array}
\usage{
\method{dim}{mlx}(x) <- value
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector that will be converted via \code{\link[=as_mlx]{as_mlx()}}.}

\item{value}{Integer vector of new dimensions}
}
\value{
Reshaped mlx object
}
\description{
Reshapes the MLX array to the specified dimensions. The total number of
elements must remain the same.
}
\examples{
x <- as_mlx(1:12)
dim(x) <- c(3, 4)
dim(x)
}
