% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{format.mlx_stream}
\alias{format.mlx_stream}
\title{Format method for mlx_stream}
\usage{
\method{format}{mlx_stream}(x, ...)
}
\arguments{
\item{x}{An mlx_stream object.}

\item{...}{Additional arguments; ignored.}
}
\value{
A character string.
}
\description{
Format method for mlx_stream
}
