% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_allofv}
\alias{get_allofv}
\title{Get objective function values of all model runs}
\usage{
get_allofv(mlxruns, sortby = c("none", "OFV", "AIC", "BIC", "BICc"))
}
\arguments{
\item{mlxruns}{\verb{<chr>} A vector containing the model file names.}

\item{sortby}{\verb{<chr>} Sort results by: "none", "OFV", "AIC", "BICc".}
}
\value{
A data frame containing the objective function values of the models.
}
\description{
Get objective function values of all model runs
}
\examples{
\dontrun{
get_allofv("r01_model.mlxtran")
}
}
