% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_fname_ofv}
\alias{get_fname_ofv}
\title{Get log-likelihood file name from Monolix results}
\usage{
get_fname_ofv(mlxrun)
}
\arguments{
\item{mlxrun}{\verb{<chr>} A Monolix file name.}
}
\value{
The file name of the population parameters of the model run.
}
\description{
Get log-likelihood file name from Monolix results
}
\examples{
\dontrun{
get_fname_ofv("r01_model.mlxtran")
}
}
