% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_fname_para}
\alias{get_fname_para}
\title{Get population parameter file name from Monolix results}
\usage{
get_fname_para(mlxrun)
}
\arguments{
\item{mlxrun}{\verb{<chr>} A Monolix file name.}
}
\value{
The file name of the population parameters of the model run.
}
\description{
Get population parameter file name from Monolix results
}
\examples{
\dontrun{
get_fname_para("r01_model.mlxtran")
}
}
