% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_fname_summ}
\alias{get_fname_summ}
\title{Get summary file name from Monolix results}
\usage{
get_fname_summ(mlxrun)
}
\arguments{
\item{mlxrun}{\verb{<chr>} A Monolix file name.}
}
\value{
The file name of the summary file of the model run.
}
\description{
Get summary file name from Monolix results
}
\examples{
\dontrun{
get_fname_summ("r01_model.mlxtran")
}
}
