% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_ofv}
\alias{get_ofv}
\title{Get OFV values from Monolix results}
\usage{
get_ofv(mlxrun)
}
\arguments{
\item{mlxrun}{\verb{<chr>} A Monolix file name.}
}
\value{
The file name of the objective function values of the model run.
}
\description{
Get OFV values from Monolix results
}
\examples{
\dontrun{
get_ofv("r01_model.mlxtran")
}
}
