% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_ofv_all}
\alias{get_ofv_all}
\title{Get objective function values of all model runs}
\usage{
get_ofv_all(
  mlxruns,
  ref = NULL,
  sortby = c("none", "OFV", "AIC", "BIC", "BICc")
)
}
\arguments{
\item{mlxruns}{\verb{<chr>} A vector containing the model file names.}

\item{ref}{\verb{<chr>} String to search for Monolix file as the reference run.}

\item{sortby}{\verb{<chr>} Sort results by: "none", "OFV", "AIC", "BICc".}
}
\value{
A data frame containing the objective function values of the models.
}
\description{
Get objective function values of all model runs
}
\examples{
\dontrun{
get_ofv_all("r01_model.mlxtran")
}
}
