% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_para}
\alias{get_para}
\title{Get parameter results of a Monolix model}
\usage{
get_para(mlxrun)
}
\arguments{
\item{mlxrun}{\verb{<chr>} A Monolix file name.}
}
\value{
A data frame containing the parameter values of the model run.
}
\description{
Get parameter results of a Monolix model
}
\examples{
\dontrun{
get_para("r01_model.mlxtran")
}
}
