% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_para_all}
\alias{get_para_all}
\title{Get parameters of all runs}
\usage{
get_para_all(mlxruns)
}
\arguments{
\item{mlxruns}{\verb{<chr>} A vector containing the model file names.}
}
\value{
All parameter values
}
\description{
Get parameters of all runs
}
\examples{
\dontrun{
get_para_all("r01_model.mlxtran")
}
}
