% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.R
\name{get_parafname}
\alias{get_parafname}
\title{Get population parameter file names of a Monolix model}
\usage{
get_parafname(mlxrun)
}
\arguments{
\item{mlxrun}{\verb{<chr>} A Monolix file name.}
}
\value{
The file name of the population parameters of the model run.
}
\description{
Get population parameter file names of a Monolix model
}
\examples{
get_parafname("r01_model.mlxtran")
}
