% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_res_ca}
\alias{get_res_ca}
\title{Read results from all Monolix runs in convergement assessment}
\usage{
get_res_ca(type = c("ofv", "pop", "saem"), path, mlxbase)
}
\arguments{
\item{type}{\verb{<chr>} Type of results to return.
Accept one of the following:
"ofv": OFV;
"pop": Population parameters;
"saem": SAEM convergence.}

\item{path}{\verb{<chr>} Path to Monolix run.}

\item{mlxbase}{\verb{<chr>} Monolix file name of the model for assessment.}
}
\value{
A data frame containing the selected results.
}
\description{
Read results from all Monolix runs in convergement assessment
}
\examples{
\dontrun{
path = "run"
get_res_ca("ofv",path,"r01_base")
}
}
