% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{get_summfname}
\alias{get_summfname}
\title{Get summary file names of a Monolix model}
\usage{
get_summfname(mlxrun)
}
\arguments{
\item{mlxrun}{\verb{<chr>} A Monolix file name.}
}
\value{
The file name of the summary file of the model run.
}
\description{
Get summary file names of a Monolix model
}
\examples{
\dontrun{
get_summfname("r01_model.mlxtran")
}
}
