% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.mlx}
\alias{is.mlx}
\title{Test if object is an MLX array}
\usage{
is.mlx(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical scalar.
}
\description{
Test if object is an MLX array
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
is.mlx(x)
}
