% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is_mlx}
\alias{is_mlx}
\title{Test if object is an MLX array}
\usage{
is_mlx(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical scalar.
}
\description{
Test if object is an MLX array
}
\examples{
x <- mlx_matrix(1:4, 2, 2)
is_mlx(x)
}
