% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\docType{methods}
\name{kronecker}
\alias{kronecker}
\alias{kronecker.default}
\alias{kronecker,mlx,mlx-method}
\alias{kronecker,mlx,ANY-method}
\alias{kronecker,ANY,mlx-method}
\alias{kronecker.mlx}
\title{Kronecker product dispatcher}
\usage{
kronecker(X, Y, FUN = "*", make.dimnames = FALSE, ...)

kronecker.default(X, Y, FUN = "*", make.dimnames = FALSE, ...)

\S4method{kronecker}{mlx,mlx}(X, Y, FUN = "*", make.dimnames = FALSE, ...)

\S4method{kronecker}{mlx,ANY}(X, Y, FUN = "*", make.dimnames = FALSE, ...)

\S4method{kronecker}{ANY,mlx}(X, Y, FUN = "*", make.dimnames = FALSE, ...)

kronecker.mlx(X, Y, FUN = "*", ..., make.dimnames = FALSE)
}
\arguments{
\item{X}{a vector or array.}

\item{Y}{a vector or array.}

\item{FUN}{Must be \code{'*'} (other functions are unsupported for MLX tensors).}

\item{make.dimnames}{logical: provide dimnames that are the product of the
   dimnames of \code{X} and \code{Y}.}

\item{...}{optional arguments to be passed to \code{FUN}.}
}
\value{
An \code{mlx} array.
}
\description{
Wrapper around \code{\link[base:kronecker]{base::kronecker()}} that enables S3 dispatch for \code{mlx} arrays
while delegating to base R for all other inputs.

Ensures the base \code{kronecker()} generic can dispatch on S3 \code{mlx} objects when
S4 dispatch is unavailable.
}
