% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlx_array_params}
\alias{mlx_array_params}
\title{Parameters for MLX array operation functions}
\arguments{
\item{x}{An \code{mlx} array or object coercible to \code{mlx} (varies by function).}

\item{device}{Target device ("gpu" or "cpu").}
}
\description{
Parameters for MLX array operation functions
}
