% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{mlx_array_required}
\alias{mlx_array_required}
\title{Parameters for Functions Requiring MLX Arrays}
\arguments{
\item{x}{An mlx array.}
}
\description{
Parameters for Functions Requiring MLX Arrays
}
\keyword{internal}
