% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlx_creation_params}
\alias{mlx_creation_params}
\title{Parameters for MLX tensor creation functions}
\arguments{
\item{dim}{Integer vector giving the tensor shape.}

\item{dtype}{MLX dtype. Supported types vary by function (see Details).}

\item{device}{Target device ("gpu" or "cpu").}
}
\description{
Parameters for MLX tensor creation functions
}
