% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg_adv.R
\name{mlx_cross}
\alias{mlx_cross}
\title{Vector cross product with MLX tensors}
\usage{
mlx_cross(a, b, axis = -1L)
}
\arguments{
\item{a, b}{Input \code{mlx} tensors containing 3D vectors.}

\item{axis}{Axis along which to compute the cross product (1-indexed, default last).}
}
\value{
An \code{mlx} tensor of cross products.
}
\description{
Vector cross product with MLX tensors
}
\examples{
u <- as_mlx(c(1, 0, 0))
v <- as_mlx(c(0, 1, 0))
mlx_cross(u, v)
}
