% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{mlx_device}
\alias{mlx_device}
\title{Get or set current MLX device}
\usage{
mlx_device(value)
}
\arguments{
\item{value}{New current device ("gpu" or "cpu"). If missing, returns the
current device.}
}
\value{
Current device (character).
}
\description{
Get or set current MLX device
}
\examples{
mlx_device()  # Get current device
mlx_device("cpu")  # Set to CPU
if (mlx_has_gpu()) {
  mlx_device("gpu")  # Set back to GPU
  mlx_device()
}
mlx_device("cpu")
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/metal.html}{mlx.core.default_device}
}
