% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlx_dtype}
\alias{mlx_dtype}
\title{Get data type helper}
\usage{
mlx_dtype(x)
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector converted via \code{\link[=as_mlx]{as_mlx()}}.}
}
\value{
Data type string
}
\description{
Get data type helper
}
\examples{
x <- as_mlx(matrix(1:6, 2, 3))
mlx_dtype(x)
}
