% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{mlx_eval}
\alias{mlx_eval}
\title{Force evaluation of an MLX operations}
\usage{
mlx_eval(x)
}
\arguments{
\item{x}{An mlx array.}
}
\value{
The input object, invisibly.
}
\description{
By default MLX computations are lazy. \code{mlx_eval(x)} forces the computations
behind \code{x} to run. You can do the same by calling (e.g.)
\link[=as.matrix.mlx]{as.matrix(x)}.
}
\examples{
system.time(x <- mlx_rand_normal(1e7))
system.time(mlx_eval(x))
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.eval}{mlx.core.eval}
}
