% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_key}
\alias{mlx_key}
\alias{mlx_key_split}
\title{Construct MLX random number generator keys}
\usage{
mlx_key(seed)

mlx_key_split(key, num = 2L)
}
\arguments{
\item{seed}{Integer or numeric seed (converted to unsigned 64-bit).}

\item{key}{An \code{mlx} key array returned by \code{\link[=mlx_key]{mlx_key()}}.}

\item{num}{Number of subkeys to produce (default 2L).}
}
\value{
An \code{mlx} array holding the PRNG key.

A list of \code{num} \code{mlx} key arrays.
}
\description{
\code{mlx_key()} provides access to MLX's stateless PRNG. Given a 64-bit seed it
returns a key that can be passed to other random helpers. Use
\code{mlx_key_split()} to derive multiple independent keys from an existing key.
}
\examples{
k <- mlx_key(42)
subkeys <- mlx_key_split(k, num = 2)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.key}
}
