% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{mlx_matrix_required}
\alias{mlx_matrix_required}
\title{Parameters for Functions Requiring MLX Matrices}
\arguments{
\item{x}{An mlx matrix (2-dimensional array).}
}
\description{
Parameters for Functions Requiring MLX Matrices
}
\keyword{internal}
