% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{mlx_maximum}
\alias{mlx_maximum}
\title{Elementwise maximum of two MLX tensors}
\usage{
mlx_maximum(x, y)
}
\arguments{
\item{x, y}{\code{mlx} tensors or objects coercible with \code{\link[=as_mlx]{as_mlx()}}.}
}
\value{
An \code{mlx} tensor containing the elementwise maximum.
}
\description{
Elementwise maximum of two MLX tensors
}
\examples{
\dontrun{
mlx_maximum(1:3, c(3, 2, 1))
}
}
