% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_param_set_values}
\alias{mlx_param_set_values}
\title{Assign tensors back to parameters}
\usage{
mlx_param_set_values(params, values)
}
\arguments{
\item{params}{A list of \code{mlx_param}.}

\item{values}{A list of tensors.}
}
\description{
Assign tensors back to parameters
}
\examples{
set.seed(1)
layer <- mlx_linear(2, 1)
params <- mlx_parameters(layer)
current <- mlx_param_values(params)
mlx_param_set_values(params, current)
}
