% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_parameters}
\alias{mlx_parameters}
\title{Collect parameters from modules}
\usage{
mlx_parameters(module)
}
\arguments{
\item{module}{An \code{mlx_module} or list of modules.}
}
\value{
A list of \code{mlx_param} objects.
}
\description{
Collect parameters from modules
}
\examples{
set.seed(1)
layer <- mlx_linear(2, 1)
mlx_parameters(layer)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/nn.html#mlx.nn.Module.parameters}
}
