% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlx_params}
\alias{mlx_params}
\title{Common MLX parameter documentation}
\arguments{
\item{x}{An \code{mlx} tensor, matrix, array, or object coercible to \code{mlx}.}

\item{dim}{Integer vector giving the tensor shape.}

\item{dtype}{Desired MLX dtype ("float32" or "float64").}

\item{device}{Target device ("gpu" or "cpu").}
}
\description{
Common MLX parameter documentation
}
