% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_bernoulli}
\alias{mlx_rand_bernoulli}
\title{Sample Bernoulli random variables on mlx arrays}
\usage{
mlx_rand_bernoulli(dim, prob = 0.5, device = mlx_default_device())
}
\arguments{
\item{dim}{Integer vector specifying array dimensions (shape).}

\item{prob}{Probability of a one.}

\item{device}{Execution target: supply \code{"gpu"}, \code{"cpu"}, or an
\code{mlx_stream} created via \code{\link[=mlx_new_stream]{mlx_new_stream()}}. By default, many
functions use the \code{\link[=mlx_device]{mlx_device()}} of their first argument.}
}
\value{
An mlx boolean array.
}
\description{
Sample Bernoulli random variables on mlx arrays
}
\examples{
mask <- mlx_rand_bernoulli(c(4, 4), prob = 0.3)
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.bernoulli}{mlx.core.random.bernoulli}
}
