% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlx_reshape}
\alias{mlx_reshape}
\title{Reshape an mlx array}
\usage{
mlx_reshape(x, newshape)
}
\arguments{
\item{x}{An mlx array.}

\item{newshape}{Integer vector specifying the new dimensions.}
}
\value{
An mlx array with the specified shape.
}
\description{
Reshape an mlx array
}
\examples{
x <- as_mlx(1:12)
mlx_reshape(x, c(3, 4))
mlx_reshape(x, c(2, 6))
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/_autosummary/mlx.core.reshape.html}
}
