% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_sequential}
\alias{mlx_sequential}
\title{Compose modules sequentially}
\usage{
mlx_sequential(...)
}
\arguments{
\item{...}{Modules to compose.}
}
\value{
An \code{mlx_module}.
}
\description{
Compose modules sequentially
}
