% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_sigmoid}
\alias{mlx_sigmoid}
\title{Sigmoid activation}
\usage{
mlx_sigmoid()
}
\value{
An \code{mlx_module} applying sigmoid activation.
}
\description{
Sigmoid activation
}
\examples{
act <- mlx_sigmoid()
x <- as_mlx(matrix(c(-2, -1, 0, 1, 2), 5, 1))
mlx_forward(act, x)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/nn.html#mlx.nn.Sigmoid}
}
