% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_silu}
\alias{mlx_silu}
\title{SiLU (Swish) activation}
\usage{
mlx_silu()
}
\value{
An \code{mlx_module} applying SiLU activation.
}
\description{
Sigmoid Linear Unit, also known as Swish activation.
}
\examples{
act <- mlx_silu()
x <- as_mlx(matrix(c(-2, -1, 0, 1, 2), 5, 1))
mlx_forward(act, x)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/nn.html#mlx.nn.SiLU}
}
