% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_tanh}
\alias{mlx_tanh}
\title{Tanh activation}
\usage{
mlx_tanh()
}
\value{
An \code{mlx_module} applying hyperbolic tangent activation.
}
\description{
Tanh activation
}
\examples{
act <- mlx_tanh()
x <- as_mlx(matrix(c(-2, -1, 0, 1, 2), 5, 1))
mlx_forward(act, x)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/nn.html#mlx.nn.Tanh}
}
