% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_mlx}
\alias{new_mlx}
\title{Internal constructor for mlx objects}
\usage{
new_mlx(ptr, dim, dtype, device)
}
\arguments{
\item{ptr}{External pointer to MLX array}

\item{dim}{Dimensions}

\item{dtype}{Data type}

\item{device}{Device}
}
\description{
Internal constructor for mlx objects
}
\keyword{internal}
