% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{Ops.mlx}
\alias{Ops.mlx}
\title{Arithmetic and comparison operators for MLX arrays}
\usage{
\method{Ops}{mlx}(e1, e2 = NULL)
}
\arguments{
\item{e1}{First operand (mlx or numeric)}

\item{e2}{Second operand (mlx or numeric)}
}
\value{
An \code{mlx} object
}
\description{
Arithmetic and comparison operators for MLX arrays
}
