% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{print.mlx_stream}
\alias{print.mlx_stream}
\title{Print method for mlx_stream}
\usage{
\method{print}{mlx_stream}(x, ...)
}
\arguments{
\item{x}{An mlx_stream object.}

\item{...}{Additional arguments; ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for mlx_stream
}
