% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmlx-package.R
\docType{package}
\name{rmlx-package}
\alias{rmlx}
\alias{rmlx-package}
\title{rmlx: Convenient Functions for Monolix Modelling}
\description{
A collection of convenient functions to facilitate modelling with Monolix. Common tasks include getting objective function values, showing parameter estimates and compiling results from multiple runs for easy comparison.
}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com} [copyright holder]

}
\keyword{internal}
