% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rowMeans.mlx}
\alias{rowMeans.mlx}
\title{Row means of MLX matrix}
\usage{
\method{rowMeans}{mlx}(x, na.rm = FALSE, dims = 1, ...)
}
\arguments{
\item{x}{An \code{mlx} matrix}

\item{na.rm}{Ignored (for compatibility)}

\item{dims}{Which dimensions are regarded as rows (default: 1)}

\item{...}{Additional arguments (ignored)}
}
\value{
An \code{mlx} vector with row means
}
\description{
Row means of MLX matrix
}
