% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_cov2023}
\alias{run_cov2023}
\title{Run covariate model building using Monolix v2023}
\usage{
run_cov2023(mlx_dir, mlx_tran, mlx_tool)
}
\arguments{
\item{mlx_dir}{\verb{<chr>} Location of the Monolix file.}

\item{mlx_tran}{\verb{<chr>} File name of path of the Monolix file.}

\item{mlx_tool}{\verb{<chr>} Tool for model building.}
}
\value{
Job ID of the run on the cluster.
}
\description{
Run covariate model building using Monolix v2023
}
\examples{
\dontrun{
run_cov2023("r01_model.mlxtran","cossac",".")
}
}
