% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_cov2024}
\alias{run_cov2024}
\title{Run covariate model building using Monolix v2024}
\usage{
run_cov2024(mlx_dir, mlx_tran, mlx_tool)
}
\arguments{
\item{mlx_dir}{\verb{<chr>} Location of the Monolix file.}

\item{mlx_tran}{\verb{<chr>} File name of path of the Monolix file.}

\item{mlx_tool}{\verb{<chr>} Tool for model building with reference to the config file.}
}
\value{
Job ID of the run on the cluster.
}
\description{
Run covariate model building using Monolix v2024
}
\examples{
\dontrun{
mlx_tran = "r01_model.mlxtran"
mlx_tool = paste0(' -U "-t modelBuilding --config r20_config_cossac.txt"')
run_cov2024(mlx_tran,mlx_tool,".")
}
}
