% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run_mlx_cmd}
\alias{run_mlx_cmd}
\title{Run Monolix model file using command line}
\usage{
run_mlx_cmd(mlx_dir, mlx_tran, wait = F)
}
\arguments{
\item{mlx_dir}{\verb{<chr>} Location of the Monolix file.}

\item{mlx_tran}{\verb{<chr>} File name of path of the Monolix file.}

\item{wait}{\verb{<lgl>} \code{TRUE} to wait for the run to finish.}
}
\value{
Job ID of the run on the cluster.
}
\description{
Run Monolix model file using command line
}
\examples{
\dontrun{
run_mlx_cmd("r01_model.mlxtran")
}
}
