% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{see_para}
\alias{see_para}
\title{Compare parameters of two model runs}
\usage{
see_para(addto = NULL, run1 = NA, run2 = NA, rse = TRUE, cv = FALSE, mlxruns)
}
\arguments{
\item{addto}{\verb{<dfr>} Output from previous run to add to if any.}

\item{run1, run2}{\verb{<chr>} Partial model names to by matched by regular expression}

\item{rse}{\verb{<lgl>} \code{TRUE} to return RSE.}

\item{cv}{\verb{<lgl>} \code{TRUE} to return CV.}

\item{mlxruns}{\verb{<chr>} A vector containing the model file names.}
}
\value{
A data frame containing all parameter values.
}
\description{
Compare parameters of two model runs
}
\examples{
\dontrun{
see_para(mlxruns=mlxruns) 
}
}
