% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{see_para_all}
\alias{see_para_all}
\title{Compare parameters of all runs}
\usage{
see_para_all(mlxruns, rse = TRUE, cv = FALSE)
}
\arguments{
\item{mlxruns}{\verb{<chr>} A vector containing the model file names.}

\item{rse}{\verb{<lgl>} \code{TRUE} to return RSE.}

\item{cv}{\verb{<lgl>} \code{TRUE} to return CV.}
}
\value{
A data frame containing the model parameters.
}
\description{
Compare parameters of all runs
}
\examples{
\dontrun{
see_para_all("r01_model.mlxtran") 
}
}
