% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[.mlx}
\alias{[.mlx}
\title{Subset MLX array}
\usage{
\method{[}{mlx}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An \code{mlx} object}

\item{i}{Row indices}

\item{j}{Column indices (for matrices)}

\item{...}{Additional indices}

\item{drop}{Should dimensions be dropped? (default: TRUE)}
}
\value{
Subsetted \code{mlx} object
}
\description{
Subset MLX array
}
