% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sum.mlx}
\alias{sum.mlx}
\title{Sum of MLX array elements}
\usage{
\method{sum}{mlx}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An \code{mlx} object}

\item{...}{Additional arguments (ignored)}

\item{na.rm}{Ignored (for compatibility)}
}
\value{
An \code{mlx} scalar
}
\description{
Sum of MLX array elements
}
