% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{t.mlx}
\alias{t.mlx}
\title{Transpose of MLX matrix}
\usage{
\method{t}{mlx}(x)
}
\arguments{
\item{x}{An \code{mlx} matrix}
}
\value{
Transposed \code{mlx} matrix
}
\description{
Transpose of MLX matrix
}
\examples{
x <- as_mlx(matrix(1:6, 2, 3))
t(x)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.transpose}
}
