% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.array.mlx}
\alias{as.array.mlx}
\title{Convert MLX array to R matrix/array}
\usage{
\method{as.array}{mlx}(x, ...)
}
\arguments{
\item{x}{An mlx array.}

\item{...}{Additional arguments (ignored)}
}
\value{
A vector, matrix or array.

A numeric array.
}
\description{
MLX vectors or scalars are returned as R vectors.
}
\examples{
Convert MLX array to R array

x <- as_mlx(matrix(1:8, 2, 4))
as.array(x)
}
