% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.matrix.mlx}
\alias{as.matrix.mlx}
\title{Convert MLX array to R matrix/array}
\usage{
\method{as.matrix}{mlx}(x, ...)
}
\arguments{
\item{x}{An \code{mlx} object}

\item{...}{Additional arguments (ignored)}
}
\value{
A numeric matrix or array
}
\description{
Convert MLX array to R matrix/array
}
