% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.matrix.mlx}
\alias{as.matrix.mlx}
\title{Convert MLX array to R matrix/array}
\usage{
\method{as.matrix}{mlx}(x, ...)
}
\arguments{
\item{x}{An mlx array.}

\item{...}{Additional arguments (ignored)}
}
\value{
A matrix or array (numeric or logical depending on dtype)
}
\description{
Convert MLX array to R matrix/array
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
as.matrix(x)
}
