% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.matrix.mlx}
\alias{as.matrix.mlx}
\title{Convert MLX array to R matrix}
\usage{
\method{as.matrix}{mlx}(x, ...)
}
\arguments{
\item{x}{An mlx array.}

\item{...}{Additional arguments (ignored)}
}
\value{
A vector, matrix or array (numeric or logical depending on dtype).
}
\description{
MLX arrays with other than 2 dimensions are converted to
a 1 column matrix, with a warning.
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
as.matrix(x)
}
