% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as.vector.mlx}
\alias{as.vector.mlx}
\title{Convert MLX array to R vector}
\usage{
\method{as.vector}{mlx}(x, mode = "any")
}
\arguments{
\item{x}{An \code{mlx} object}

\item{mode}{Character string specifying the mode (ignored)}
}
\value{
A numeric vector
}
\description{
Convert MLX array to R vector
}
