% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as_mlx}
\alias{as_mlx}
\title{Create MLX array from R object}
\usage{
as_mlx(x, dtype = c("float32", "float64"), device = mlx_default_device())
}
\arguments{
\item{x}{Numeric vector, matrix, or array to convert}

\item{dtype}{Data type: "float32" or "float64" (default)}

\item{device}{Device: "gpu" (default) or "cpu"}
}
\value{
An object of class \code{mlx}
}
\description{
Create MLX array from R object
}
\examples{
\dontrun{
x <- as_mlx(matrix(1:12, 3, 4))
}
}
