% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{crossprod.mlx}
\alias{crossprod.mlx}
\title{Cross product}
\usage{
\method{crossprod}{mlx}(x, y = NULL, ...)
}
\arguments{
\item{x}{An \code{mlx} matrix}

\item{y}{An \code{mlx} matrix (default: NULL, uses x)}

\item{...}{Additional arguments passed to base::crossprod.}
}
\value{
\code{t(x) \%*\% y} as an \code{mlx} object
}
\description{
Cross product
}
\examples{
x <- as_mlx(matrix(1:6, 2, 3))
crossprod(x)
}
