% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.mlx}
\alias{dim.mlx}
\title{Get dimensions of MLX array}
\usage{
\method{dim}{mlx}(x)
}
\arguments{
\item{x}{An \code{mlx} object}
}
\value{
Integer vector of dimensions
}
\description{
Get dimensions of MLX array
}
