% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dim.mlx}
\alias{dim.mlx}
\title{Get dimensions of MLX array}
\usage{
\method{dim}{mlx}(x)
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector that will be converted via \code{\link[=as_mlx]{as_mlx()}}.}
}
\value{
Integer vector of dimensions
}
\description{
Get dimensions of MLX array
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
dim(x)
}
