% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{exam_runs}
\alias{exam_runs}
\title{Examine objective function and parameter values of selected runs}
\usage{
exam_runs(runnums, path = ".", ifOFV = TRUE, ifParam = TRUE)
}
\arguments{
\item{runnums}{\verb{<chr>} Partial model file names of the model files.}

\item{path}{\verb{<chr>} Path to model directory for \link{get_mlx}.}

\item{ifOFV}{\verb{<lgl>} \code{TRUE} to return objective function values.}

\item{ifParam}{\verb{<lgl>} \code{TRUE} to return parameter values.}
}
\value{
A list containing the OFV and parameter values of the selected models.
}
\description{
Examine objective function and parameter values of selected runs
}
\examples{
\dontrun{
# Model files are matched by regular expression 
exam_runs("r01|r02") # model names include "r01" and "r02"
exam_runs("/r01|/r02") # model names include "/r01" and "/r02"
exam_runs("^r01|^r02") # model names starting with "r01" and "r02"
}
}
