% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.R
\name{get_allofv}
\alias{get_allofv}
\title{Get objective function values of all model runs}
\usage{
get_allofv(mlxruns, sortAIC = FALSE)
}
\arguments{
\item{mlxruns}{\verb{<chr>} A vector containing the model file names.}

\item{sortAIC}{\verb{<lgl>} \code{TRUE} to sort results by AIC values.}
}
\value{
A data frame containing the objective function values of the models.
}
\description{
Get objective function values of all model runs
}
\examples{
\dontrun{
get_allofv("r01_model.mlxtran")
}
}
