% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.R
\name{get_mlx}
\alias{get_mlx}
\title{Search for Monolix model files}
\usage{
get_mlx(path = ".", ifall = TRUE)
}
\arguments{
\item{path}{\verb{<chr>} Path to the directory of the model files.}

\item{ifall}{\verb{<lgl>} \code{TRUE} to return all model files in the directory.}
}
\value{
A character vector containing the names of the model files.
}
\description{
Search for Monolix model files
}
\examples{
\dontrun{
mlx_path = "CLOU064C1/mas/mas_1/model/pgm_001/Task_02_REMODEL_Sim_PopPK_plan/runs" 
get_mlx(mlx_path)
}
}
