% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx.R
\name{get_ofv}
\alias{get_ofv}
\title{Get summary file names of a Monolix model}
\usage{
get_ofv(mlxrun)
}
\arguments{
\item{mlxrun}{\verb{<chr>} A Monolix file name.}
}
\value{
The file name of the objective function values of the model run.
}
\description{
Get summary file names of a Monolix model
}
\examples{
\dontrun{
get_ofv("r01_model.mlxtran")
}
}
