% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{\%*\%.mlx}
\alias{\%*\%.mlx}
\title{Matrix multiplication for MLX arrays}
\usage{
\method{\%*\%}{mlx}(x, y)
}
\arguments{
\item{x, y}{numeric or complex matrices or vectors.}
}
\value{
An \code{mlx} object
}
\description{
Matrix multiplication for MLX arrays
}
