% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{\%*\%.mlx}
\alias{\%*\%.mlx}
\title{Matrix multiplication for MLX arrays}
\usage{
\method{\%*\%}{mlx}(x, y)
}
\arguments{
\item{x}{First matrix (mlx or numeric)}

\item{y}{Second matrix (mlx or numeric)}
}
\value{
An \code{mlx} object
}
\description{
Matrix multiplication for MLX arrays
}
