% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{\%*\%.mlx}
\alias{\%*\%.mlx}
\title{Matrix multiplication for MLX arrays}
\usage{
\method{\%*\%}{mlx}(x, y)
}
\arguments{
\item{x, y}{numeric or complex matrices or vectors.}
}
\value{
An mlx object
}
\description{
Matrix multiplication for MLX arrays
}
\examples{
\dontrun{
x <- as_mlx(matrix(1:6, 2, 3))
y <- as_mlx(matrix(1:6, 3, 2))
x \%*\% y
}
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.matmul}
}
