% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{length.mlx}
\alias{length.mlx}
\title{Get length of MLX array}
\usage{
\method{length}{mlx}(x)
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector that will be converted via \code{\link[=as_mlx]{as_mlx()}}.}
}
\value{
Total number of elements.
}
\description{
Get length of MLX array
}
\examples{
x <- as_mlx(matrix(1:6, 2, 3))
length(x)
}
