% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{Math.mlx}
\alias{Math.mlx}
\title{Math operations for MLX arrays}
\usage{
\method{Math}{mlx}(x, ...)
}
\arguments{
\item{x}{An \code{mlx} object}

\item{...}{Additional arguments (ignored)}
}
\value{
An \code{mlx} object with the result
}
\description{
Math operations for MLX arrays
}
