% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{Math.mlx}
\alias{Math.mlx}
\title{Math operations for MLX arrays}
\usage{
\method{Math}{mlx}(x, ...)
}
\arguments{
\item{x}{An \code{mlx} object}

\item{...}{Additional arguments (ignored)}
}
\value{
An \code{mlx} object with the result
}
\description{
Math operations for MLX arrays
}
\examples{
x <- as_mlx(matrix(c(-1, 0, 1), 3, 1))
sin(x)
round(x + 0.4)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/array.html}
}
