% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mean.mlx}
\alias{mean.mlx}
\title{Mean of MLX array elements}
\usage{
\method{mean}{mlx}(x, ...)
}
\arguments{
\item{x}{An \code{mlx} object}

\item{...}{Additional arguments (ignored)}
}
\value{
An \code{mlx} scalar
}
\description{
Mean of MLX array elements
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
mean(x)
}
