% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimnames.R
\name{mlx-dimnames}
\alias{mlx-dimnames}
\alias{dimnames.mlx}
\alias{dimnames<-.mlx}
\alias{names.mlx}
\alias{names<-.mlx}
\title{Dimnames and names for MLX arrays}
\usage{
\method{dimnames}{mlx}(x)

\method{dimnames}{mlx}(x) <- value

\method{names}{mlx}(x)

\method{names}{mlx}(x) <- value
}
\arguments{
\item{x}{An object.}

\item{value}{Replacement names or dimnames.}
}
\value{
The requested names, or \code{x} with updated metadata for replacement
forms.
}
\description{
Since version 0.4.0, Rmlx supports character names and dimnames. These work
much like base R: \code{\link[=names]{names()}}, \code{\link[=rownames]{rownames()}}, \code{\link[=colnames]{colnames()}}, \code{\link[=dimnames]{dimnames()}} and
associated setters all work. Dimnames may be \code{NULL} as a whole, and any
individual dimension's names may be \code{NULL}.
}
\details{
Dimnames can be convenient, but they also slow down operations by adding work
in base R (mlx can't store character strings). In particular, slowdowns
for subsetting can be considerable, maybe 5x or 6x slower. To maximize
performance, you can avoid names, or remove them with \code{\link[=unname]{unname()}}.
}
