% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlx_array_params}
\alias{mlx_array_params}
\title{Roxygen parameters for arrays}
\arguments{
\item{x}{An \code{mlx} array or object coercible to \code{mlx} (varies by function).}

\item{device}{Target device ("gpu" or "cpu").}
}
\description{
Roxygen parameters for arrays
}
\keyword{internal}
