% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{mlx_contiguous}
\alias{mlx_contiguous}
\title{Ensure contiguous memory layout}
\usage{
mlx_contiguous(x)
}
\arguments{
\item{x}{An mlx array.}
}
\value{
An mlx array backed by contiguous storage.
}
\description{
Returns a copy of \code{x} with contiguous strides.
}
\examples{
x <- mlx_swapaxes(mlx_matrix(1:4, 2, 2), axis1 = 1, axis2 = 2)
y <- mlx_contiguous(x)
identical(as.array(x), as.array(y))
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.contiguous}
}
