% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{mlx_default_device}
\alias{mlx_default_device}
\title{Get or set default MLX device}
\usage{
mlx_default_device(value)
}
\arguments{
\item{value}{New default device ("gpu" or "cpu"). If missing, returns current default.}
}
\value{
Current default device (character)
}
\description{
Get or set default MLX device
}
\examples{
mlx_default_device()  # Get current default
mlx_default_device("cpu")  # Set to CPU
mlx_default_device("gpu")  # Set back to GPU
mlx_default_device()
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/metal.html}{mlx.core.default_device}
}
