% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{mlx_device}
\alias{mlx_device}
\title{Get device associated with an MLX object}
\usage{
mlx_device(x)
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector that will be converted via \code{\link[=as_mlx]{as_mlx()}}.}
}
\value{
\code{"gpu"} or \code{"cpu"}.
}
\description{
Get device associated with an MLX object
}
\examples{
x <- as_mlx(1:10)
mlx_device(x)
}
