% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlx_dim}
\alias{mlx_dim}
\title{Get dimensions helper}
\usage{
mlx_dim(x)
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector that will be converted via \code{\link[=as_mlx]{as_mlx()}}.}
}
\value{
Dimensions
}
\description{
Get dimensions helper
}
\examples{
x <- as_mlx(matrix(1:6, 2, 3))
mlx_dim(x)
}
