% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mlx_dtype}
\alias{mlx_dtype}
\title{Get the data type of an MLX array}
\usage{
mlx_dtype(x)
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector that will be converted via \code{\link[=as_mlx]{as_mlx()}}.}
}
\value{
A data type string (see \code{\link[=as_mlx]{as_mlx()}} for possibilities).
}
\description{
Get the data type of an MLX array
}
\examples{
x <- mlx_matrix(1:6, 2, 3)
mlx_dtype(x)
}
