% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg_adv.R
\name{mlx_eig}
\alias{mlx_eig}
\title{Eigen decomposition for MLX tensors}
\usage{
mlx_eig(x)
}
\arguments{
\item{x}{An \code{mlx} square matrix.}
}
\value{
A list with components \code{values} and \code{vectors}, both \code{mlx} tensors.
}
\description{
Eigen decomposition for MLX tensors
}
\examples{
x <- as_mlx(matrix(c(2, -1, 0, 2), 2, 2))
eig <- mlx_eig(x)
eig$values
eig$vectors
}
