% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg_adv.R
\name{mlx_eigh}
\alias{mlx_eigh}
\title{Eigen decomposition of Hermitian MLX tensors}
\usage{
mlx_eigh(x, uplo = c("L", "U"))
}
\arguments{
\item{x}{An \code{mlx} square matrix.}

\item{uplo}{Character string indicating which triangle to use ("L" or "U").}
}
\value{
A list with components \code{values} and \code{vectors}.
}
\description{
Eigen decomposition of Hermitian MLX tensors
}
\examples{
x <- as_mlx(matrix(c(2, 1, 1, 3), 2, 2))
mlx_eigh(x)
}
