% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg_adv.R
\name{mlx_eigvals}
\alias{mlx_eigvals}
\title{Eigenvalues of mlx arrays}
\usage{
mlx_eigvals(x)
}
\arguments{
\item{x}{An mlx matrix (2-dimensional array).}
}
\value{
An mlx array containing eigenvalues.
}
\description{
Eigenvalues of mlx arrays
}
\examples{
x <- as_mlx(matrix(c(3, 1, 0, 2), 2, 2))
mlx_eigvals(x)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/linalg.html#mlx.linalg.eigvals}
}
