% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg_adv.R
\name{mlx_eigvalsh}
\alias{mlx_eigvalsh}
\title{Eigenvalues of Hermitian MLX tensors}
\usage{
mlx_eigvalsh(x, uplo = c("L", "U"))
}
\arguments{
\item{x}{An \code{mlx} square matrix.}

\item{uplo}{Character string indicating which triangle to use ("L" or "U").}
}
\value{
An \code{mlx} tensor containing eigenvalues.
}
\description{
Eigenvalues of Hermitian MLX tensors
}
\examples{
x <- as_mlx(matrix(c(2, 1, 1, 3), 2, 2))
mlx_eigvalsh(x)
}
