% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{mlx_eval}
\alias{mlx_eval}
\title{Force evaluation of lazy MLX operations}
\usage{
mlx_eval(x)
}
\arguments{
\item{x}{An \code{mlx} object}
}
\value{
The input object (invisibly)
}
\description{
Force evaluation of lazy MLX operations
}
