% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_forward}
\alias{mlx_forward}
\title{Forward pass utility}
\usage{
mlx_forward(module, x)
}
\arguments{
\item{module}{An \code{mlx_module}.}

\item{x}{An mlx array.}
}
\value{
Output array.
}
\description{
Forward pass utility
}
\examples{
set.seed(1)
layer <- mlx_linear(2, 1)
input <- as_mlx(matrix(c(1, 2), 1, 2))
mlx_forward(layer, input)
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/nn.html#mlx.nn.Module}{mlx.nn.Module}
}
