% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{mlx_identity}
\alias{mlx_identity}
\title{Identity matrices on MLX devices}
\usage{
mlx_identity(n, dtype = c("float32", "float64"), device = mlx_default_device())
}
\arguments{
\item{n}{Size of the square matrix.}

\item{dtype}{MLX dtype to use. One of \code{"float32"}, \code{"float64"}, \code{"int8"},
\code{"int16"}, \code{"int32"}, \code{"int64"}, \code{"uint8"}, \code{"uint16"}, \code{"uint32"},
\code{"uint64"}, \code{"bool"}, or \code{"complex64"}.}

\item{device}{Character string specifying the device for computation.
Options: \code{"gpu"} (default) or \code{"cpu"}. Default: \code{mlx_default_device()}.}
}
\value{
An mlx identity matrix.
}
\description{
Identity matrices on MLX devices
}
\examples{
I4 <- mlx_identity(4)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.identity}
}
