% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_linear}
\alias{mlx_linear}
\title{Create a learnable linear transformation}
\usage{
mlx_linear(
  in_features,
  out_features,
  bias = TRUE,
  device = mlx_default_device()
)
}
\arguments{
\item{in_features}{Number of input features.}

\item{out_features}{Number of output features.}

\item{bias}{Should a bias term be included?}

\item{device}{Device for the parameters.}
}
\value{
An object of class \code{mlx_module}.
}
\description{
Create a learnable linear transformation
}
\examples{
set.seed(1)
layer <- mlx_linear(3, 2)
x <- as_mlx(matrix(1:6, 2, 3))
mlx_forward(layer, x)
}
