% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mlx_load_gguf}
\alias{mlx_load_gguf}
\title{Load MLX tensors from the GGUF format}
\usage{
mlx_load_gguf(file, device = mlx_default_device())
}
\arguments{
\item{file}{Path to a \code{.npy} file. The extension is appended automatically
when missing.}

\item{device}{Target device for the loaded tensor (\code{"gpu"} or \code{"cpu"}).}
}
\value{
A list containing:
\describe{
\item{\code{tensors}}{Named list of \code{mlx} arrays.}
\item{\code{metadata}}{Named list where values are \code{NULL}, character vectors, or
\code{mlx} arrays depending on the GGUF entry type.}
}
}
\description{
Load MLX tensors from the GGUF format
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/io.html#mlx.core.load_gguf}
}
