% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mlx_load_safetensors}
\alias{mlx_load_safetensors}
\title{Load MLX tensors from the safetensors format}
\usage{
mlx_load_safetensors(file, device = mlx_default_device())
}
\arguments{
\item{file}{Path to a \code{.npy} file. The extension is appended automatically
when missing.}

\item{device}{Target device for the loaded tensor (\code{"gpu"} or \code{"cpu"}).}
}
\value{
A list containing:
\describe{
\item{\code{tensors}}{Named list of \code{mlx} arrays.}
\item{\code{metadata}}{Named character vector with the serialized metadata.}
}
}
\description{
Load MLX tensors from the safetensors format
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/io.html#mlx.core.load_safetensors}
}
