% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{mlx_logcumsumexp}
\alias{mlx_logcumsumexp}
\title{Log cumulative sum exponential for mlx arrays}
\usage{
mlx_logcumsumexp(x, axis = NULL, reverse = FALSE, inclusive = TRUE)
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector that will be converted via \code{\link[=as_mlx]{as_mlx()}}.}

\item{axis}{Optional axis (single integer) to operate over.}

\item{reverse}{Logical flag for reverse accumulation.}

\item{inclusive}{Logical flag controlling inclusivity.}
}
\value{
An mlx array.
}
\description{
Log cumulative sum exponential for mlx arrays
}
\examples{
x <- as_mlx(1:4)
as.vector(as.matrix(mlx_logcumsumexp(x)))
m <- as_mlx(matrix(1:6, 2, 3))
as.matrix(mlx_logcumsumexp(m, axis = 2))
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.logaddexp}
}
