% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{mlx_maximum}
\alias{mlx_maximum}
\title{Elementwise maximum of two mlx arrays}
\usage{
mlx_maximum(x, y)
}
\arguments{
\item{x, y}{mlx arrays or objects coercible with \code{\link[=as_mlx]{as_mlx()}}.}
}
\value{
An mlx array containing the elementwise maximum.
}
\description{
Elementwise maximum of two mlx arrays
}
\examples{
\dontrun{
mlx_maximum(1:3, c(3, 2, 1))
}
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.maximum}
}
