% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{mlx_ones}
\alias{mlx_ones}
\title{Create arrays of ones on MLX devices}
\usage{
mlx_ones(
  dim,
  dtype = c("float32", "float64", "int8", "int16", "int32", "int64", "uint8", "uint16",
    "uint32", "uint64", "bool", "complex64")
)
}
\arguments{
\item{dim}{Integer vector specifying array dimensions (shape).}

\item{dtype}{Data type string. Supported types include:
\itemize{
\item Floating point: \code{"float32"}, \code{"float64"}
\item Integer: \code{"int8"}, \code{"int16"}, \code{"int32"}, \code{"int64"}, \code{"uint8"}, \code{"uint16"}, \code{"uint32"}, \code{"uint64"}
\item Other: \code{"bool"}, \code{"complex64"}
}

Not all functions support all types. See individual function documentation.}
}
\value{
An mlx array filled with ones.
}
\description{
Create arrays of ones on MLX devices
}
\examples{
ones <- with_device("cpu", mlx_ones(c(2, 2), dtype = "float64"))
ones_int <- mlx_ones(c(3, 3), dtype = "int32")
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.ones}{mlx.core.ones}
}
