% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{mlx_ones}
\alias{mlx_ones}
\title{Create arrays of ones on MLX devices}
\usage{
mlx_ones(
  dim,
  dtype = c("float32", "float64", "int8", "int16", "int32", "int64", "uint8", "uint16",
    "uint32", "uint64", "bool", "complex64"),
  device = mlx_default_device()
)
}
\arguments{
\item{dim}{Integer vector specifying the array shape/dimensions.}

\item{dtype}{MLX dtype to use. One of \code{"float32"}, \code{"float64"}, \code{"int8"},
\code{"int16"}, \code{"int32"}, \code{"int64"}, \code{"uint8"}, \code{"uint16"}, \code{"uint32"},
\code{"uint64"}, \code{"bool"}, or \code{"complex64"}.}

\item{device}{Execution target: provide \code{"gpu"}, \code{"cpu"}, or an
\code{mlx_stream} created via \code{\link[=mlx_new_stream]{mlx_new_stream()}}. Defaults to the current
\code{\link[=mlx_default_device]{mlx_default_device()}}.}
}
\value{
An mlx array filled with ones.
}
\description{
Create arrays of ones on MLX devices
}
\examples{
ones <- mlx_ones(c(2, 2), dtype = "float64", device = "cpu")
ones_int <- mlx_ones(c(3, 3), dtype = "int32")
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.ones}{mlx.core.ones}
}
