% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_param_values}
\alias{mlx_param_values}
\title{Retrieve parameter tensors}
\usage{
mlx_param_values(params)
}
\arguments{
\item{params}{A list of \code{mlx_param}.}
}
\value{
List of \code{mlx} tensors.
}
\description{
Retrieve parameter tensors
}
\examples{
set.seed(1)
layer <- mlx_linear(2, 1)
vals <- mlx_param_values(mlx_parameters(layer))
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/nn.html#mlx.nn.Module.parameters}
}
