% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_bernoulli}
\alias{mlx_rand_bernoulli}
\title{Sample Bernoulli random variables on mlx arrays}
\usage{
mlx_rand_bernoulli(dim, prob = 0.5, device = mlx_default_device())
}
\arguments{
\item{dim}{Integer vector specifying the array shape/dimensions.}

\item{prob}{Probability of a one.}

\item{device}{Character string specifying the device for computation.
Options: \code{"gpu"} (default) or \code{"cpu"}. Default: \code{mlx_default_device()}.}
}
\value{
An mlx boolean array.
}
\description{
Sample Bernoulli random variables on mlx arrays
}
\examples{
mask <- mlx_rand_bernoulli(c(4, 4), prob = 0.3)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.bernoulli}
}
