% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_bernoulli}
\alias{mlx_rand_bernoulli}
\title{Sample Bernoulli random variables on MLX tensors}
\usage{
mlx_rand_bernoulli(dim, prob = 0.5, device = mlx_default_device())
}
\arguments{
\item{dim}{Integer vector giving the tensor shape.}

\item{prob}{Probability of a one.}

\item{device}{Target device ("gpu" or "cpu").}
}
\value{
An \code{mlx} boolean tensor.
}
\description{
Sample Bernoulli random variables on MLX tensors
}
