% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_gumbel}
\alias{mlx_rand_gumbel}
\title{Sample from the Gumbel distribution on mlx arrays}
\usage{
mlx_rand_gumbel(dim, dtype = c("float32", "float64"))
}
\arguments{
\item{dim}{Integer vector specifying array dimensions (shape).}

\item{dtype}{Data type string. Supported types include:
\itemize{
\item Floating point: \code{"float32"}, \code{"float64"}
\item Integer: \code{"int8"}, \code{"int16"}, \code{"int32"}, \code{"int64"}, \code{"uint8"}, \code{"uint16"}, \code{"uint32"}, \code{"uint64"}
\item Other: \code{"bool"}, \code{"complex64"}
}

Not all functions support all types. See individual function documentation.}
}
\value{
An mlx array with Gumbel-distributed entries.
}
\description{
Sample from the Gumbel distribution on mlx arrays
}
\examples{
samples <- mlx_rand_gumbel(c(2, 3))
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.gumbel}{mlx.core.random.gumbel}
}
