% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_gumbel}
\alias{mlx_rand_gumbel}
\title{Sample from the Gumbel distribution on mlx arrays}
\usage{
mlx_rand_gumbel(
  dim,
  dtype = c("float32", "float64"),
  device = mlx_default_device()
)
}
\arguments{
\item{dim}{Integer vector specifying the array shape/dimensions.}

\item{dtype}{Desired MLX dtype ("float32" or "float64").}

\item{device}{Character string specifying the device for computation.
Options: \code{"gpu"} (default) or \code{"cpu"}. Default: \code{mlx_default_device()}.}
}
\value{
An mlx array with Gumbel-distributed entries.
}
\description{
Sample from the Gumbel distribution on mlx arrays
}
\examples{
samples <- mlx_rand_gumbel(c(2, 3))
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.gumbel}
}
