% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_normal}
\alias{mlx_rand_normal}
\title{Sample from a normal distribution on mlx arrays}
\usage{
mlx_rand_normal(dim, mean = 0, sd = 1, dtype = c("float32", "float64"))
}
\arguments{
\item{dim}{Integer vector specifying array dimensions (shape).}

\item{mean}{Mean of the normal distribution.}

\item{sd}{Standard deviation of the normal distribution.}

\item{dtype}{Data type string. Supported types include:
\itemize{
\item Floating point: \code{"float32"}, \code{"float64"}
\item Integer: \code{"int8"}, \code{"int16"}, \code{"int32"}, \code{"int64"}, \code{"uint8"}, \code{"uint16"}, \code{"uint32"}, \code{"uint64"}
\item Other: \code{"bool"}, \code{"complex64"}
}

Not all functions support all types. See individual function documentation.}
}
\value{
An mlx array with normally distributed entries.
}
\description{
Sample from a normal distribution on mlx arrays
}
\examples{
weights <- mlx_rand_normal(c(3, 3), mean = 0, sd = 0.1)
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.normal}{mlx.core.random.normal}
}
