% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_randint}
\alias{mlx_rand_randint}
\title{Sample random integers on MLX tensors}
\usage{
mlx_rand_randint(
  dim,
  low,
  high,
  dtype = c("int32", "int64", "uint32", "uint64"),
  device = mlx_default_device()
)
}
\arguments{
\item{dim}{Integer vector giving the tensor shape.}

\item{low}{Lower bound (inclusive).}

\item{high}{Upper bound (exclusive).}

\item{dtype}{Desired integer dtype ("int32", "int64", "uint32", "uint64").}

\item{device}{Target device ("gpu" or "cpu").}
}
\value{
An \code{mlx} tensor of random integers.
}
\description{
Generates random integers uniformly distributed over the interval [low, high).
}
\examples{
# Random integers from 0 to 9
samples <- mlx_rand_randint(c(3, 3), low = 0, high = 10)

# Random integers from -5 to 4
samples <- mlx_rand_randint(c(2, 5), low = -5, high = 5)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.randint}
}
