% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_truncated_normal}
\alias{mlx_rand_truncated_normal}
\title{Sample from a truncated normal distribution on mlx arrays}
\usage{
mlx_rand_truncated_normal(lower, upper, dim, dtype = c("float32", "float64"))
}
\arguments{
\item{lower}{Lower bound of the truncated normal.}

\item{upper}{Upper bound of the truncated normal.}

\item{dim}{Integer vector specifying array dimensions (shape).}

\item{dtype}{Data type string. Supported types include:
\itemize{
\item Floating point: \code{"float32"}, \code{"float64"}
\item Integer: \code{"int8"}, \code{"int16"}, \code{"int32"}, \code{"int64"}, \code{"uint8"}, \code{"uint16"}, \code{"uint32"}, \code{"uint64"}
\item Other: \code{"bool"}, \code{"complex64"}
}

Not all functions support all types. See individual function documentation.}
}
\value{
An mlx array with truncated normally distributed entries.
}
\description{
Sample from a truncated normal distribution on mlx arrays
}
\examples{
samples <- mlx_rand_truncated_normal(-1, 1, c(5, 5))
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.truncated_normal}{mlx.core.random.truncated_normal}
}
