% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_truncated_normal}
\alias{mlx_rand_truncated_normal}
\title{Sample from a truncated normal distribution on MLX tensors}
\usage{
mlx_rand_truncated_normal(
  lower,
  upper,
  dim,
  dtype = c("float32", "float64"),
  device = mlx_default_device()
)
}
\arguments{
\item{lower}{Lower bound of the truncated normal.}

\item{upper}{Upper bound of the truncated normal.}

\item{dim}{Integer vector giving the tensor shape.}

\item{dtype}{Desired MLX dtype ("float32" or "float64").}

\item{device}{Target device ("gpu" or "cpu").}
}
\value{
An \code{mlx} tensor with truncated normally distributed entries.
}
\description{
Sample from a truncated normal distribution on MLX tensors
}
\examples{
samples <- mlx_rand_truncated_normal(-1, 1, c(5, 5))
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.truncated_normal}
}
