% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_truncated_normal}
\alias{mlx_rand_truncated_normal}
\title{Sample from a truncated normal distribution on mlx arrays}
\usage{
mlx_rand_truncated_normal(
  lower,
  upper,
  dim,
  dtype = c("float32", "float64"),
  device = mlx_default_device()
)
}
\arguments{
\item{lower}{Lower bound of the truncated normal.}

\item{upper}{Upper bound of the truncated normal.}

\item{dim}{Integer vector specifying the array shape/dimensions.}

\item{dtype}{Desired MLX dtype ("float32" or "float64").}

\item{device}{Execution target: provide \code{"gpu"}, \code{"cpu"}, or an
\code{mlx_stream} created via \code{\link[=mlx_new_stream]{mlx_new_stream()}}. Defaults to the current
\code{\link[=mlx_default_device]{mlx_default_device()}}.}
}
\value{
An mlx array with truncated normally distributed entries.
}
\description{
Sample from a truncated normal distribution on mlx arrays
}
\examples{
samples <- mlx_rand_truncated_normal(-1, 1, c(5, 5))
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.truncated_normal}{mlx.core.random.truncated_normal}
}
