% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{mlx_rand_uniform}
\alias{mlx_rand_uniform}
\title{Sample from a uniform distribution on mlx arrays}
\usage{
mlx_rand_uniform(
  dim,
  min = 0,
  max = 1,
  dtype = c("float32", "float64"),
  device = mlx_default_device()
)
}
\arguments{
\item{dim}{Integer vector specifying the array shape/dimensions.}

\item{min}{Lower bound for the uniform distribution.}

\item{max}{Upper bound for the uniform distribution.}

\item{dtype}{Desired MLX dtype ("float32" or "float64").}

\item{device}{Character string specifying the device for computation.
Options: \code{"gpu"} (default) or \code{"cpu"}. Default: \code{mlx_default_device()}.}
}
\value{
An mlx array whose entries are sampled uniformly.
}
\description{
Sample from a uniform distribution on mlx arrays
}
\examples{
noise <- mlx_rand_uniform(c(2, 2), min = -1, max = 1)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/random.html#mlx.core.random.uniform}
}
