% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{mlx_softmax_layer}
\alias{mlx_softmax_layer}
\title{Softmax activation}
\usage{
mlx_softmax_layer(axis = -1L)
}
\arguments{
\item{axis}{Axis along which to apply softmax (default: -1, last axis).}
}
\value{
An \code{mlx_module} applying softmax activation.
}
\description{
Softmax activation
}
\examples{
act <- mlx_softmax_layer()
x <- as_mlx(matrix(1:6, 2, 3))
mlx_forward(act, x)
}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/nn.html#mlx.nn.Softmax}{mlx.nn.Softmax}
}
