% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{mlx_synchronize}
\alias{mlx_synchronize}
\title{Synchronize MLX device}
\usage{
mlx_synchronize(device = c("gpu", "cpu"))
}
\description{
Waits for all outstanding operations on the specified device to complete.
}
