% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{mlx_tile}
\alias{mlx_tile}
\title{Tile a tensor}
\usage{
mlx_tile(x, reps)
}
\arguments{
\item{x}{An \code{mlx} tensor.}

\item{reps}{Integer vector giving the number of repetitions for each axis.}
}
\value{
An \code{mlx} tensor with tiled content.
}
\description{
Tile a tensor
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
mlx_tile(x, reps = c(1, 2))
}
