% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{mlx_train_step}
\alias{mlx_train_step}
\title{Single training step helper}
\usage{
mlx_train_step(module, loss_fn, optimizer, ...)
}
\arguments{
\item{module}{An \code{mlx_module}.}

\item{loss_fn}{Function of \code{module} and data returning an \code{mlx} scalar.}

\item{optimizer}{Optimizer object from \code{mlx_optimizer_sgd()}.}

\item{...}{Additional data passed to \code{loss_fn}.}
}
\value{
A list with the current loss.
}
\description{
Single training step helper
}
