% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{mlx_zeros}
\alias{mlx_zeros}
\title{Create arrays of zeros on MLX devices}
\usage{
mlx_zeros(dim, dtype = c("float32", "float64"), device = mlx_default_device())
}
\arguments{
\item{dim}{Integer vector specifying the array shape/dimensions.}

\item{dtype}{MLX dtype to use (\code{"float32"} or \code{"float64"}).}

\item{device}{Character string specifying the device for computation.
Options: \code{"gpu"} (default) or \code{"cpu"}. Default: \code{mlx_default_device()}.}
}
\value{
An mlx array filled with zeros.
}
\description{
Create arrays of zeros on MLX devices
}
\examples{
zeros <- mlx_zeros(c(2, 3))
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/array.html#mlx.core.zeros}
}
