% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mlx}
\alias{print.mlx}
\title{Print MLX array}
\usage{
\method{print}{mlx}(x, ...)
}
\arguments{
\item{x}{An mlx array, or an R array/matrix/vector that will be converted via \code{\link[=as_mlx]{as_mlx()}}.}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{x}, invisibly.
}
\description{
Printing an array only evaluates it if it is of small size (less than
100 elements and 2 dimensions)
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
print(x)
}
