% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mlx}
\alias{print.mlx}
\title{Print MLX array}
\usage{
\method{print}{mlx}(x, ...)
}
\arguments{
\item{x}{An \code{mlx} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print MLX array
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
print(x)
}
