% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxmlx-package.R
\docType{package}
\name{rmlx-package}
\alias{rmlx}
\alias{rmlx-package}
\title{rmlx: Convenient Functions for Processing of Monolix Results}
\description{
A collection of functions to facilitate processing of Monolix results.
}
\author{
\strong{Maintainer}: Tomas Sou \email{tomas.sou@carexer.com} [copyright holder]

}
\keyword{internal}
