% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res.R
\name{see_runs}
\alias{see_runs}
\title{See summary of run results as kables}
\usage{
see_runs(runnums, path = ".", ...)
}
\arguments{
\item{runnums}{\verb{<chr>} Starting string of Monolix run files separated by '|'.}

\item{path}{\verb{<chr>} Location of Monolix run files.}

\item{...}{Additional arguments for \code{\link[=exam_runs]{exam_runs()}}}
}
\value{
A list containing the OFV and parameter values of the selected models.
}
\description{
See summary of run results as kables
}
\examples{
\dontrun{
# Model files are matched by regular expression 
see_runs("r01|r02") # model names include "r01" and "r02"
see_runs("/r01|/r02") # model names include "/r01" and "/r02"
see_runs("^r01|^r02") # model names starting with "r01" and "r02"
}
}
