% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str.mlx}
\alias{str.mlx}
\title{Object structure for MLX array}
\usage{
\method{str}{mlx}(object, ...)
}
\arguments{
\item{object}{An \code{mlx} object}

\item{...}{Additional arguments (ignored)}
}
\description{
Object structure for MLX array
}
