% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str.mlx}
\alias{str.mlx}
\title{Object structure for MLX array}
\usage{
\method{str}{mlx}(object, ...)
}
\arguments{
\item{object}{An mlx object}

\item{...}{Additional arguments (ignored)}
}
\value{
\code{NULL} invisibly.
}
\description{
Object structure for MLX array
}
\examples{
x <- as_mlx(matrix(1:4, 2, 2))
str(x)
}
