% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tcrossprod.mlx}
\alias{tcrossprod.mlx}
\title{Transposed cross product}
\usage{
\method{tcrossprod}{mlx}(x, y = NULL)
}
\arguments{
\item{x}{An \code{mlx} matrix}

\item{y}{An \code{mlx} matrix (default: NULL, uses x)}
}
\value{
\code{x \%*\% t(y)} as an \code{mlx} object
}
\description{
Transposed cross product
}
