% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{with_default_device}
\alias{with_default_device}
\title{Temporarily set the default MLX device}
\usage{
with_default_device(device = c("gpu", "cpu"), code)
}
\arguments{
\item{device}{Device to use (\code{"gpu"} or \code{"cpu"}).}

\item{code}{Expression to evaluate while \code{device} is active.}
}
\value{
The result of evaluating \code{code}.
}
\description{
Temporarily set the default MLX device
}
\examples{
old <- mlx_default_device()
with_default_device("cpu", mlx_default_device())
mlx_default_device(old)
}
\seealso{
\url{https://ml-explore.github.io/mlx/build/html/python/metal.html}
}
