% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{with_default_device}
\alias{with_default_device}
\title{Temporarily set the default MLX device}
\usage{
with_default_device(device, code)
}
\arguments{
\item{device}{\code{"gpu"}, \code{"cpu"}, or an \code{mlx_stream} created via \code{\link[=mlx_new_stream]{mlx_new_stream()}}.}

\item{code}{Expression to evaluate while \code{device} is active.}
}
\value{
The result of evaluating \code{code}.
}
\description{
Temporarily set the default MLX device
}
\examples{
with_default_device("cpu", x <- mlx_vector(1:10))

}
\seealso{
\href{https://ml-explore.github.io/mlx/build/html/python/metal.html}{mlx.core.default_device}
}
